dependencies: avoid canonicalization for missing files
Checks whether a file exists before canonicalizing. This is a significant speed
up because canonicalizing walks the path from the root, which is much more
likely to spend time randomly stat'ing files that end up not being relevant.

This more than doubles (!) the speed of the program on my machine.