dependencies: cache file existence when canonicalizing
After caching canonicalization, the calls to std::fs::exists were making up a
large portion of the remaining syscalls. Caching these, too, yields another
30-40% speedup.

After this commit, on large projects, the hash table for the cache itself
dominates performance measurements, which is about as far as I'm willing to go
for now.